kmswebrtcbaseconnection: add get_certificate_pem
erdtlssrtpdec from Ericson already generates a certificate so if we
want to use a custom one we must provide it as a string containing
a X509 certificate and RSA private key in PEM format. So far we use
the one generated by erdtlssrtpdec and provide the method to get it.

Change-Id: Ia83cdf2769c7adcea588342c26b9f16ae1060338